COMPLETE: A relatively straightforward implementation of the Monologued server.
This is obviously a version 1.0, but it taught me a few things about Go's
networking and its batteries included philosophy.  I'll have a longer blog
post about it in a few days.  It still needs a configuration file and a
couple of other features found in traditional finger servers, but for now
this is a pretty good example.